Change the CullingData to be a CPU buffer
As when we remove a model bundle we can't just change the total count, as that makes the amount of models from the end to not render. So, we have to go through every single one, even though it has been removed. But processing removed model is useless, so we can change the modelCount to 0 and so the compute shader will skip that model.