Remove start barriers
As we can't set the previous state to undefined. And it is difficult to remember the previous state, if you don't update the barrier info on every use. So, the new approach would be create the texture in the correct state and revert the state back if it is changed for some reason.